+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 2. [](#improved)
     * Added optional `$scope` parameter to `ObjectController::checkAuthorization()`
 3. [](#bugfix)
-    * Fixed continue with `PageInterface` types
+    * Fixed continue task with `PageInterface` types
     
 # v1.1.6
 ## 11/29/2021