+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # v1.1.7
-## mm/dd/2021
+## 01/03/2022
 
 1. [](#new)
     * Allow intercepting object `create`, `update` and `delete` tasks by using `FlexTaskEvent` event